Structure du projet

+++ src/caractere.h
@@ -3,7 +3,7 @@
 #include <Arduino.h>
 
 struct Caractere {
-    uint8_t code;      // Code ASCII du glyphe
+    uint16_t code;      // Code ASCII du glyphe
     uint8_t bitmap[8]; // Bitmap du glyphe (8x8)
 };
